refactor(Toolbar)[OK-89] Create Control folder for Controls in Toolbar
